applied clippy suggestions and removed outdated 'clippy.toml'

+++ src/ddnnf/anomalies/t_wise_sampling/t_iterator.rs
@@ -21,11 +21,11 @@ impl StreamingIterator for TIndicesIter {
         }
 
         self.tuple[0] += 1;
-        if self.tuple[0] as usize >= self.number_of_vars {
+        if self.tuple[0] >= self.number_of_vars {
             let mut p: usize = 0;
 
             // carry over to the next places - like 0999 -> 1000
-            while self.tuple[p] as usize >= self.number_of_vars - p
+            while self.tuple[p] >= self.number_of_vars - p
                 && self.tuple[self.t] == 0
             {
                 self.tuple[p] = 0;
